project/requirements.txt
torch
torchvision
Pillow
transformers
tqdm
scikit-learn
numpy
torch-summary